extract constants in one file
