--- discord-groups/constants.js
@@ -1,4 +1,0 @@
-const NO_SPACES_ALLOWED = 'Roles cannot have spaces';
-const CANNOT_CONTAIN_GROUP = "Roles cannot contain 'group'.";
-
-export { NO_SPACES_ALLOWED, CANNOT_CONTAIN_GROUP };
