--- extension-requests/constants.js
@@ -1,19 +1,0 @@
-const taskInfoModelHeadings = [
-  { title: 'Title' },
-  { title: 'Ends On', key: 'endsOn', time: true },
-  { title: 'Purpose' },
-  { title: 'Assignee' },
-  { title: 'Created By', key: 'createdBy' },
-  { title: 'Is Noteworthy', key: 'isNoteworthy' },
-];
-
-const extensionRequestCardHeadings = [
-  { title: 'Title' },
-  { title: 'Reason' },
-  { title: 'Old Ends On', key: 'oldEndsOn', time: true },
-  { title: 'New Ends On', key: 'newEndsOn', time: true },
-  { title: 'Status', bold: true },
-  { title: 'Assignee' },
-  { title: 'Created At', key: 'timestamp', time: true },
-  { title: 'Task', key: 'taskId' },
-];
