+++ extension-requests/index.html
@@ -102,7 +102,6 @@
     </div>
     <h2 id="error"></h2>
     <script src="/constants.js"></script>
-    <script src="constants.js"></script>
     <script src="/utils.js"></script>
     <script src="local-utils.js"></script>
     <script src="script.js"></script>
