+++ online-members/online-members.html
@@ -10,7 +10,6 @@
     <link rel="stylesheet" href="/style.css" async />
     <link rel="stylesheet" href="/online-members/online-members.css" />
     <script src="/constants.js"></script>
-    <script src="/online-members/constants.js" defer></script>
     <script src="/online-members/sse-events.js" defer></script>
     <script src="/online-members/utils.js" defer></script>
     <script src="/online-members/script.js" defer></script>
