--- profile/constants.js
@@ -1,13 +1,0 @@
-export const YEARS_OF_EXPERIENCE = 'yoe';
-export const SUPER_USER = 'super_user';
-export const APPROVE_BUTTON_TEXT = 'Approve';
-export const REJECT_BUTTON_TEXT = 'Reject';
-export const APPROVAL_PROMPT_TEXT = 'Reason for Approval';
-export const ALERT_APPROVED = 'User Data Approved !!!';
-export const ALERT_ERROR = 'Something went wrong. Please check console errors.';
-export const ALERT_REJECTED = 'User Data Rejected!!!';
-export const OLD_DATA = 'old-data';
-export const NEW_DATA = 'new-data';
-export const DIFF_CLASS = 'diff';
-export const OLD_DIFF_CLASS = 'oldDiff';
-export const NEW_DIFF_CLASS = 'newDiff';
