--- users/constants.js
@@ -1,22 +1,0 @@
-const RDS_API_USERS = `${API_BASE_URL}/users`;
-const RDS_API_SKILLS = `${API_BASE_URL}/tags`;
-const USER_LIST_ELEMENT = 'user-list';
-const LOADER_ELEMENT = 'loader';
-const TILE_VIEW_BTN = 'tile-view-btn';
-const TABLE_VIEW_BTN = 'table-view-btn';
-const USER_SEARCH_ELEMENT = 'user-search';
-const DEFAULT_AVATAR = '/users/images/avatar.png';
-const PAGINATION_ELEMENT = 'pagination';
-const PREV_BUTTON = 'prevButton';
-const NEXT_BUTTON = 'nextButton';
-const USER_FETCH_COUNT = 100;
-const NONE = 'NONE';
-const OOO = 'OOO';
-const ACTIVE = 'ACTIVE';
-const IDLE = 'IDLE';
-const FILTER_MODAL = 'filter-modal';
-const FILTER_BUTTON = 'filter-button';
-const AVAILABILITY_FILTER = 'filter-by-avaviability';
-const APPLY_FILTER_BUTTON = 'apply-filter-button';
-const CLEAR_BUTTON = 'clear-button';
-const STATUS_LIST = ['ooo', 'active', 'idle'];
